+++ src/io/writer/mod.rs
@@ -1,19 +1,20 @@
 pub mod stdout;
 pub mod temp_file;
 
+use crate::io::writer::stdout::StdoutWriter;
 use crate::io::writer::temp_file::TempFile;
 use anyhow::Result;
 
 pub enum Writer {
     TempFile(TempFile),
-    Stdout,
+    Stdout(StdoutWriter),
 }
 
 impl Writer {
-    pub fn write(&mut self, line: &str) -> Result<()> {
+    pub fn write_batch<'a>(&mut self, lines: impl Iterator<Item = &'a str>) -> Result<()> {
         match self {
-            Writer::TempFile(w) => w.write(line),
-            Writer::Stdout => stdout::write_line(line),
+            Writer::TempFile(w) => w.write_batch(lines),
+            Writer::Stdout(w) => w.write_batch(lines),
         }
     }
 }